Fix a goroutine leak and simplify the code
In general, it is dangerous to send to a unbufferred channel without selecting on another channel, e.g. context.Done(). For example, in the modified code, a goroutine that sends a message to connectToACS would block forever if the main loop was stopped
because context was canceled. 

Fix the problem simply by using a buffered channel. Since the channel is now buffered, we no longer need to spin up a new goroutine to send a message to the channel.